Red rectangle can be moved around. Animations are coming next.

+++ src/entity/mob/ai.hpp
@@ -4,7 +4,7 @@ class Mob;
 class AI {
 public:
     AI(Mob* mob);
-    virtual void Update(float delta) = 0;
+    virtual void Update(double delta) = 0;
 
 protected:
     Mob* mob_;
